add thconcept base model class to ensure same logic
